requirements.txt
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.0.3
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
networkx==2.3
numpy==1.16.4
Pillow==6.0.0
pyparsing==2.4.0
pypng==0.0.19
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
Werkzeug==0.15.4
